refactor: wire Coder workspace defaults into runtimes settings
- Add runtime option defaults utility for reading/writing per-runtime
  persisted defaults (SSH coder config, mode memory).
- Move Coder controls into shared runtime component and wire runtimes
  settings to shared Coder form.
- Draft-first precedence: local draft config takes priority over
  persisted defaults for responsive UI during async template/preset
  loading.
- Interaction-gated persistence: Coder config is only persisted after
  explicit user interaction, preventing silent writes on initial render.
- Decouple settings Coder defaults from mode memory: settings edits
  update only coderConfig via writeSshCoderDefaultsPreservingMode,
  preserving coderEnabled owned by creation-flow runtime selection.
- Guard Coder config writes against stale project scope switches
  using ref-based scope validation in async callbacks.
- Hide Coder config form when CLI is unavailable.
